+++ src/modules/ingredient/ingredient.service.spec.ts
@@ -1,11 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { LoggerService } from '../logger/logger.service';
 import { IngredientService } from './ingredient.service';
-import { RedisService } from '../redis/redis.service';
+import { IngredientType } from './ingredient.types';
 
 describe('IngredientService', () => {
     let ingredientService: IngredientService;
-    let redisService: RedisService;
+
+    const loggerServiceProvider = {
+        info: jest.fn(),
+        error: jest.fn()
+    };
 
     beforeEach(async () => {
         const module: TestingModule = await Test.createTestingModule({
@@ -13,27 +17,25 @@ describe('IngredientService', () => {
                 IngredientService,
                 {
                     provide: LoggerService,
-                    useValue: {
-                        info: () => {},
-                        error: () => {}
-                    }
-                },
-                {
-                    provide: RedisService,
-                    useValue: {
-                        get: jest.fn(),
-                        set: jest.fn()
-                    }
+                    useValue: loggerServiceProvider
                 }
-            ],
+            ]
         }).compile();
 
         ingredientService = module.get(IngredientService);
-        redisService = module.get(RedisService);
+        ingredientService.onModuleInit();
     });
 
     it('should be defined', () => {
         expect(ingredientService).toBeDefined();
-        expect(redisService).toBeDefined();
+    });
+
+    it('should filter ingredients when provided not supported ones', () => {
+        const mockProvidedIngredients: IngredientType[] = ['focaccia', 'bagel', 'UNKNOWN', 'XXX'];
+        const mockFilteredIngredients: IngredientType[] = ['focaccia', 'bagel'];
+
+        const filteredIngredients: IngredientType[] = ingredientService.filterIngredients(mockProvidedIngredients);
+
+        expect(filteredIngredients).toEqual(mockFilteredIngredients);
     });
 });