+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.25.0] - 2024-02-15
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Filtering provided ingredients - only supported ingredients.
+- [Marek Kurańda](https://github.com/mjkuranda): Test for `IngredientService`.
 
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): Ingredient as a string instead of object.